+++ sm64_random_assets/music_realizations/human_joncrall/__init__.py
@@ -1,0 +1,3 @@
+from .zero import generate_zero_binary
+
+__all__ = ['generate_zero_binary']