+++ sm64_random_assets/music_realizations/openai_gpt_5_6_thinking/__init__.py
@@ -1,0 +1,3 @@
+from .simple_music import generate_simple_music
+
+__all__ = ['generate_simple_music']